requirements.txt
asgiref==3.5.2
Django==3.2.15
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==4.12.0
install==1.3.5
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.3
zipp==3.8.1
